Rename the Victor 9000 decoder.
